Fixed a little thing--not really a bug.
It was only a bug in the sense that it bugged me a little. I didn't like
that I was removing an item from a list in the middle of a 'for x in l'
over the list, so I just made a removal list and then did a filter after
the fact.